Revert "v1.0.10.1 (#204)"
This reverts commit c7d359bb1917158437af5d6d211faef30c8bbe9c.
