+++ src/components/Community/post/CommentsSection.tsx
@@ -70,10 +70,7 @@ const CommentsSection = ({ apiBoardType, postId }: CommentsSectionProps) => {
                     <label className="content">{comment.content}</label>
 
                     <label className="date">
-                      {formatDateTime(comment.updatedAt)}{' '}
-                      <span className="update">
-                        {comment.isUpdated && '(수정됨)'}
-                      </span>
+                      {formatDateTime(comment.createdAt)}
                     </label>
                   </div>
                 </div>
@@ -224,11 +221,6 @@ const Comments = styled.div`
   }
 
   .date {
-    font-size: ${({ theme }) => theme.typography.fontSize.body3};
-  }
-
-  .update {
-    color: ${({ theme }) => theme.colors.gray600};
     font-size: ${({ theme }) => theme.typography.fontSize.body3};
   }
 `;
