+++ src/types/Community/comment.ts
@@ -6,9 +6,7 @@ export interface Comment {
   commentId: number;
   content: string;
   createdAt: string;
-  updatedAt: string;
   author: AuthorInfo;
-  isUpdated: boolean;
   isMyComment: boolean;
 }
